+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 优化响应服务的行为。
+  - com.dwarfeng.familyhelper.webapi.impl.service.system.PexpResponseServiceImpl。
+
 - 新建 Controller。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.NotifyHistoryController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.NotifyInfoRecordController。